requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
dataclasses==0.6
Django==4.2.15
django-cors-headers==4.4.0
django-mongoengine==0.5.6
djangorestframework==3.15.2
djongo==1.2.31
dnspython==2.6.1
idna==3.7
mongoengine==0.28.2
pyconfig==3.2.3
pydub==0.25.1
pymongo==4.8.0
python-decouple==3.8
pytool==5.0.0
pytz==2024.1
requests==2.32.3
simplejson==3.19.3
six==1.16.0
SpeechRecognition==3.10.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2